[fix] background 이미지 레이아웃 변경, 색 추가
백그라운드 이미지가 너무 커서 레이아웃이 맞지 않는 문제 해결.
배경 색 추가.
